+++ src/pyproject2conda/_typing.py
@@ -19,7 +19,11 @@ if TYPE_CHECKING:
     )
     from .utils import _Missing  # pyright: ignore[reportPrivateUsage]  # noqa: F401
 
-MISSING_TYPE: TypeAlias = "Literal[_Missing.MISSING]"
+MISSING_TYPE: TypeAlias = "Literal[_Missing.MISSING]"  # pyre-ignore[type-alias-error]
 
-RequirementCommentPair: TypeAlias = "tuple[Requirement | None, str | None]"
-RequirementOverridePair: TypeAlias = "tuple[Requirement | None, OverrideDeps | None]"
+RequirementCommentPair: TypeAlias = (
+    "tuple[Requirement | None, str | None]"  # pyre-ignore[type-alias-error]
+)
+RequirementOverridePair: TypeAlias = (
+    "tuple[Requirement | None, OverrideDeps | None]"  # pyre-ignore[type-alias-error]
+)