+++ src/pyproject2conda/_typing_compat.py
@@ -1,16 +1,5 @@
 import sys
-
-if sys.version_info >= (3, 10):
-    from typing import TypeAlias
-else:
-    from typing_extensions import TypeAlias
-
-
-if sys.version_info >= (3, 9):
-    from typing import Annotated
-else:
-    from typing_extensions import Annotated
-
+from typing import TypeAlias
 
 if sys.version_info >= (3, 11):
     from typing import Self
@@ -19,7 +8,6 @@ else:
 
 
 __all__ = [
-    "Annotated",
     "Self",
     "TypeAlias",
 ]